Did extensive code cleanup (fixing ALL compiler warnings for GCC), more robust object deletion and removal from containers, ABS and DIFF macros for absolute value and difference between 2 numbers, new common.h header file for useful stuff that is even useful in other header files (added to Code::Blocks project, Visual Studio project, and Makefile), updated lawreference.cpp in dev folder to have Elite Liberal Victory entries, updated liberalagenda.cpp to display correct Stalinist results if(won==-2), and updated AUTHORS to have full list of contributors.

+++ src/basemode/baseactions.cpp
@@ -339,7 +339,7 @@ void stopevil(void)
 
    // Start at the city level, rather than the absolute top
    if(multipleCityMode) {
-      for(l=0;l<location.size();l++) {
+      for(l=0;l<(int)location.size();l++) {
          if(location[l]->type == squad_location->city) {
             loc = l;
             break;
@@ -349,15 +349,15 @@ void stopevil(void)
 
    vector<long> temploc;
    // 1. LCS safe houses
-   for(l=0;l<location.size();l++)
+   for(l=0;l<(int)location.size();l++)
       if(location[l]->parent==loc && location[l]->renting>=0 && !location[l]->hidden)
          temploc.push_back(l);
    // 2. CCS safe houses
-   for(l=0;l<location.size();l++)
+   for(l=0;l<(int)location.size();l++)
       if(location[l]->parent==loc && location[l]->renting==RENTING_CCS && !location[l]->hidden)
          temploc.push_back(l);
    // 3. Other sites
-   for(l=0;l<location.size();l++)
+   for(l=0;l<(int)location.size();l++)
       if(location[l]->parent==loc && location[l]->renting==RENTING_NOCONTROL && !location[l]->hidden)
          temploc.push_back(l);
 
@@ -418,22 +418,22 @@ void stopevil(void)
       }*/
 
       temploc.clear();
-      for(l=0;l<location.size();l++)
+      for(l=0;l<(int)location.size();l++)
       {
          if(location[l]->parent==loc&&location[l]->renting>=0&&!location[l]->hidden)temploc.push_back(l);
       }
-      for(l=0;l<location.size();l++)
+      for(l=0;l<(int)location.size();l++)
       {
          if(location[l]->parent==loc&&location[l]->renting==RENTING_CCS&&!location[l]->hidden)temploc.push_back(l);
       }
-      for(l=0;l<location.size();l++)
+      for(l=0;l<(int)location.size();l++)
       {
          if(location[l]->parent==loc&&location[l]->renting==RENTING_NOCONTROL&&!location[l]->hidden)temploc.push_back(l);
       }
 
 
       int y=10;
-      for(p=page*11;p<temploc.size()&&p<page*11+11;p++)
+      for(p=page*11;p<(int)temploc.size()&&p<page*11+11;p++)
       {
          if(p == -1) break;
          Location* this_location = location[temploc[p]];
@@ -538,7 +538,7 @@ void stopevil(void)
          addprevpagestr();
       }
       //PAGE DOWN
-      if((page+1)*11<temploc.size())
+      if((page+1)*11<(int)temploc.size())
       {
          move(20,60);
          addnextpagestr();
@@ -557,12 +557,12 @@ void stopevil(void)
       //PAGE UP
       if((c==interface_pgup||c==KEY_UP||c==KEY_LEFT)&&page>0)page--;
       //PAGE DOWN
-      if((c==interface_pgdn||c==KEY_DOWN||c==KEY_RIGHT)&&(page+1)*11<temploc.size())page++;
+      if((c==interface_pgdn||c==KEY_DOWN||c==KEY_RIGHT)&&(page+1)*11<(int)temploc.size())page++;
 
       if(c>='a'&&c<='k')
       {
          int sq=page*11+(int)(c-'a');
-         if(sq<temploc.size()&&sq>=0)
+         if(sq<(int)temploc.size()&&sq>=0)
          {
             int oldloc=loc;
             loc=temploc[sq];
@@ -572,7 +572,7 @@ void stopevil(void)
             }
             int subcount = 0;
 
-            for(l=0;l<location.size();l++)
+            for(l=0;l<(int)location.size();l++)
                if(location[l]->parent==loc)
                   subcount++;
 
@@ -972,7 +972,7 @@ void setvehicles(void)
       int x=1,y=10;
       char str[200];
 
-      for(l=page*18;l<vehicle.size()&&l<page*18+18;l++)
+      for(l=page*18;l<(int)vehicle.size()&&l<page*18+18;l++)
       {
          set_color(COLOR_WHITE,COLOR_BLACK,0);
          for(p=0;p<6;p++)
@@ -985,7 +985,7 @@ void setvehicles(void)
                break;
             }
          }
-         for(p=0;p<pool.size();p++)
+         for(p=0;p<(int)pool.size();p++)
          {
             if(pool[p]->squadid!=-1&&pool[p]->alive&&
                pool[p]->pref_carid==vehicle[l]->id()&&pool[p]->squadid!=activesquad->id)
@@ -1018,7 +1018,7 @@ void setvehicles(void)
          addprevpagestr();
       }
       //PAGE DOWN
-      if((page+1)*18<vehicle.size())
+      if((page+1)*18<(int)vehicle.size())
       {
          move(17,53);
          addnextpagestr();
@@ -1049,7 +1049,7 @@ void setvehicles(void)
       {
          int slot=c-'A'+page*18;
 
-         if(slot>=0&&slot<vehicle.size())
+         if(slot>=0&&slot<(int)vehicle.size())
          {
             bool choice = true;
             if (activesquad->squad[0])
@@ -1100,7 +1100,7 @@ void setvehicles(void)
       {
          int slot=c-'a'+page*18;
 
-         if(slot>=0&&slot<vehicle.size())
+         if(slot>=0&&slot<(int)vehicle.size())
          {
             bool choice = true;
             if (activesquad->squad[0])
@@ -1167,7 +1167,7 @@ void setvehicles(void)
       //PAGE UP
       if((c==interface_pgup||c==KEY_UP||c==KEY_LEFT)&&page>0)page--;
       //PAGE DOWN
-      if((c==interface_pgdn||c==KEY_DOWN||c==KEY_RIGHT)&&(page+1)*18<vehicle.size())page++;
+      if((c==interface_pgdn||c==KEY_DOWN||c==KEY_RIGHT)&&(page+1)*18<(int)vehicle.size())page++;
 
    }while(1);
 }
